Fixed the matrix type
